Fix tests - add comments
